#0: Move distributed to its own directory
